requirements.txt
aiogram==3.23.0
openai==2.14.0
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0
gspread==6.2.1
oauth2client==4.1.3
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
APScheduler==3.11.2
deepdiff==8.6.1
pytz==2025.2
aiohttp==3.13.2
httpx==0.28.1
requests==2.32.5
cryptography==46.0.3
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
pydantic[email]==2.9.0
python-multipart==0.0.21
